+++ CHANGELOG.rst
@@ -1,3 +1,12 @@
+2.4.0 (2026-05-04)
+==================
+
+- Add ``http`` alias for ``HTTPServer``
+- Add ``ssl`` alias for ``SelfSignedSSLContext``
+- Cleanup code
+- Update tests
+- Update docs
+
 2.3.1 (2026-05-01)
 ==================